Fixed 'resolve common deps' spec style: moved to 'unordered' folder, removed redundant 'describe('resolve')...', file renamed to common.spec.js
