--- spec/resolve/unordered_dependencies/entity-entity.spec.js
@@ -1,170 +1,0 @@
-var expect  = require('chai').expect,
-    _       = require('lodash'),
-    resolve = require('../../../lib/index').resolve;
-
-describe('resolve', function () {
-    describe('unordered dependencies: entity - entity', function () {
-        it('should resolve entity depending on another entity', function () {
-            var decl = [{ block: 'A' }],
-                deps = [
-                    {
-                        entity: { block: 'A' },
-                        dependOn: [{ entity: { block: 'B' } }]
-                    }
-                ],
-                resolved = resolve(decl, deps);
-
-            expect(resolved.entities).to.contain({ block: 'B' });
-        });
-
-        it('should resolve entity depending on multiple entities', function () {
-            var decl = [{ block: 'A' }],
-                deps = [
-                    {
-                        entity: { block: 'A' },
-                        dependOn: [
-                            { entity: { block: 'B' } },
-                            { entity: { block: 'C' } }
-                        ]
-                    }
-                ],
-                resolved = resolve(decl, deps);
-
-            expect(resolved.entities).to.contain({ block: 'B' })
-                .and.to.contain({ block: 'C' });
-        });
-
-        it('should include entity once if entity depends on itself', function () {
-            var decl = [{ block: 'A' }],
-                deps = [
-                    {
-                        entity: { block: 'A' },
-                        dependOn: [
-                            { entity: { block: 'A' } }
-                        ]
-                    }
-                ],
-                resolved = resolve(decl, deps);
-
-            expect(resolved.entities).to.be.eql([{ block: 'A' }]);
-        });
-
-        it('should resolve dependency depending on another entity', function () {
-            var decl = [{ block: 'A' }],
-                deps = [
-                    {
-                        entity: { block: 'A' },
-                        dependOn: [
-                            { entity: { block: 'B' } }
-                        ]
-                    },
-                    {
-                        entity: { block: 'B' },
-                        dependOn: [
-                            { entity: { block: 'C' } }
-                        ]
-                    }
-                ],
-                resolved = resolve(decl, deps);
-
-            expect(resolved.entities).to.contain({ block: 'C' });
-        });
-
-        it('should allow dependency to depend on multiple another entities', function () {
-            var decl = [{ block: 'A' }],
-                deps = [
-                    {
-                        entity: { block: 'A' },
-                        dependOn: [
-                            { entity: { block: 'B' } }
-                        ]
-                    },
-                    {
-                        entity: { block: 'B' },
-                        dependOn: [
-                            { entity: { block: 'C' } },
-                            { entity: { block: 'D' } }
-                        ]
-                    }
-                ],
-                resolved = resolve(decl, deps);
-
-            expect(resolved.entities).to.contain({ block: 'C' })
-                .and.to.contain({ block: 'D' });
-        });
-
-        it('should include entity to result once if multiple entities depend on this entity', function () {
-            var decl = [
-                    { block: 'A' },
-                    { block: 'B' }
-                ],
-                deps = [
-                    {
-                        entity: { block: 'A' },
-                        dependOn: [
-                            { entity: { block: 'C' } }
-                        ]
-                    },
-                    {
-                        entity: { block: 'B' },
-                        dependOn: [
-                            { entity: { block: 'C' } }
-                        ]
-                    }
-                ],
-                resolved = resolve(decl, deps),
-                firstIndex = _.findIndex(resolved.entities, { block: 'C' }),
-                lastIndex = _.findLastIndex(resolved.entities, { block: 'C' });
-
-            expect(resolved.entities).to.contain({ block: 'C' });
-            expect(firstIndex).to.be.equal(lastIndex);
-        });
-
-        it('should not include entity if no entity from decl depends on it', function () {
-            var decl = [{ block: 'A' }],
-                deps = [
-                    {
-                        entity: { block: 'B' },
-                        dependOn: []
-                    }
-                ],
-                resolved = resolve(decl, deps);
-
-            expect(resolved.entities).to.not.contain({ block: 'B' });
-        });
-
-        it('should not include entity if no entity from decl depends on it and this entity has dependency on entity' +
-            ' listed in decl', function () {
-            var decl = [{ block: 'A' }],
-                deps = [
-                    {
-                        entity: { block: 'B' },
-                        dependOn: [
-                            { entity: { block: 'A' } }
-                        ]
-                    }
-                ],
-                resolved = resolve(decl, deps);
-
-            expect(resolved.entities).not.to.contain({ block: 'B' });
-        });
-
-        it('should not include dependencie\'s dependency if no entity from decl\'s dependencies depends ' +
-            'on it', function () {
-            var decl = [{ block: 'A' }],
-                deps = [
-                    {
-                        entity: { block: 'C' },
-                        dependOn: [
-                            {
-                                entity: { block: 'D' }
-                            }
-                        ]
-                    }
-                ],
-                resolved = resolve(decl, deps);
-
-            expect(resolved.entities).to.not.contain({ block: 'D' });
-        });
-    });
-});
